Make unclaim-bosh-lite-resource task idempotent
- This task will now succeed and result in an unclaimed environment if
  the environment lock exists and fail out if the lock resource does not
  exist

[#129169949]

+++ tasks/unclaim-bosh-lite-resource/run.sh
@@ -8,9 +8,19 @@ set -x
 
 pushd resource-pools
   echo "Unclaiming $RESOURCE_NAME"
-  git mv "$RESOURCE_TYPE/claimed/$RESOURCE_NAME" "$RESOURCE_TYPE/unclaimed/"
-  git add "$RESOURCE_TYPE"
-  git commit -m "Unclaim $RESOURCE_NAME via $PIPELINE_NAME pipeline"
+
+  if [[ -f "$RESOURCE_TYPE/claimed/$RESOURCE_NAME" || -f "$RESOURCE_TYPE/unclaimed/$RESOURCE_NAME" ]]; then
+    if [[ -f "$RESOURCE_TYPE/claimed/$RESOURCE_NAME" ]] ; then
+      git mv "$RESOURCE_TYPE/claimed/$RESOURCE_NAME" "$RESOURCE_TYPE/unclaimed/"
+      git add "$RESOURCE_TYPE"
+      git commit -m "Unclaim $RESOURCE_NAME via $PIPELINE_NAME pipeline"
+    elif [[ -f "$RESOURCE_TYPE/unclaimed/$RESOURCE_NAME" ]] ; then
+      echo "$RESOURCE_NAME is already unclaimed"
+    fi
+  else
+    echo "$RESOURCE_NAME does not currently exist in pool"
+    exit 1
+  fi
 popd
 
 rsync -a resource-pools/ resource-pools-artifacts
